+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Development
 
+- Catch Exceptions raised by a failing `text_traceback()` during error handling for CSW communication.
+- More F-Strings.
+
 ## [1.4.1](https://github.com/berlinonline/ckanext-fisbroker/releases/tag/1.4.1)
 
 _(2024-01-30)_